requirements.txt
gitpython==3.1.24  # Adjust the version as per your needs
argparse==1.4.0    # This is part of the standard library, but added here for clarity